# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Socious-API",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "https://dev.socious.io/api/v2/"
    },
    {
      "url": "http://localhost:8370"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "paths": {
    "/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "email": "{{user_login}}",
                  "password": "{{user_password}}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "first_name": "Ehsan",
                  "last_name": "Mahmoudi",
                  "username": "socious2022",
                  "password": "socious2022",
                  "email": "info_2022@socious.io"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "RefreshToken",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "example": "{{refresh_token}}"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/auth/preregister": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "preregister",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "username": "socious",
                  "email": "info@socious.io"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "Vary": {
                "schema": {
                  "type": "string",
                  "example": "Origin"
                }
              },
              "Content-Type": {
                "schema": {
                  "type": "string",
                  "example": "application/json; charset=utf-8"
                }
              },
              "Content-Length": {
                "schema": {
                  "type": "integer",
                  "example": "30"
                }
              },
              "Date": {
                "schema": {
                  "type": "string",
                  "example": "Fri, 19 Aug 2022 17:20:05 GMT"
                }
              },
              "Connection": {
                "schema": {
                  "type": "string",
                  "example": "keep-alive"
                }
              },
              "Keep-Alive": {
                "schema": {
                  "type": "string",
                  "example": "timeout=5"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "examples": {
                  "example-0": {
                    "summary": "preregister (all ok)",
                    "value": {
                      "username": null,
                      "email": null
                    }
                  },
                  "example-1": {
                    "summary": "preregister (email only)",
                    "value": {
                      "email": null
                    }
                  },
                  "example-2": {
                    "summary": "preregister (both invalid)",
                    "value": {
                      "username": "\"username\" with value \"socious team\" fails to match the required pattern: /^(?=.{6,20}$)(?![_.])(?!.*[_.]{2})[a-zA-Z0-9._]+(?<![_.])$/",
                      "email": "\"email\" must be a valid email"
                    }
                  },
                  "example-3": {
                    "summary": "preregister (mix of invalid and existing)",
                    "value": {
                      "username": "\"username\" with value \"socious team\" fails to match the required pattern: /^(?=.{6,20}$)(?![_.])(?!.*[_.]{2})[a-zA-Z0-9._]+(?<![_.])$/",
                      "email": "EXISTS"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/otp": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "send otp",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "email": "{{user_login}}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/auth/otp/confirm": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "confirm otp",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "example": "207364"
          },
          {
            "name": "email",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "example": "info@socious.io"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/auth/forget-password": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "forget password",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "email": "{{user_login}}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "refresh token",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "refresh_token": "{{refresh_token}}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
...
    "/applicants/{applicant_id}/offer": {
      "parameters": [
        {
          "name": "applicant_id",
          "in": "path",
          "schema": {
            "type": "string"
          },
          "required": true
        }
      ],
      "post": {
        "tags": [
          "project"
        ],
        "summary": "project applicatons",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "offer_text": "test offer text",
                  "offer_rate": "500$",
                  "due_date": "2022-08-12T11:38:03.40",
                  "assignment_total": 5
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/applicants/{applicant_id}": {
      "parameters": [
        {
          "name": "applicant_id",
          "in": "path",
          "schema": {
            "type": "string"
          },
          "required": true
        }
      ],
      "get": {
        "tags": [
          "project"
        ],
        "summary": "get applicant",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
...
  }
}